+++ src/api/requests/index.ts
@@ -1,3 +1,3 @@
 export * from './transaction'
 export * from './user'
-export * from './GetServerConfig'
+export * from './serverConfig'
