+++ src/api/types/serverConfig.ts
